Trabalhamos em correão de bugs

+++ app/controllers/AdmHomeController.php
@@ -5,19 +5,23 @@ require_once 'app/models/AvisosAdmModel.php';
 
 class AdmHomeController {
     public function index() {
+        // Verifica se o usuário administrador está em sessão
+        if (!isset($_SESSION['admin_id'])) {
+            header('Location: /admlogin'); // Redireciona para a página de login
+            exit();
+        }
+
         $avisos = [];
         $resumo = [
             'totalVendas' => 0,
             'valorMovimentado' => 0,
         ];
     
-        if (isset($_SESSION['admin_id'])) {
-            $admModel = new AvisosAdmModel();
-            $avisos = $admModel->getAvisos($_SESSION['admin_id']);
+        $admModel = new AvisosAdmModel();
+        $avisos = $admModel->getAvisos($_SESSION['admin_id']);
     
-            // Obter o total de vendas e o valor movimentado
-            $resumo = $this->getResumoVendas();
-        }
+        // Obter o total de vendas e o valor movimentado
+        $resumo = $this->getResumoVendas();
 
         include 'app/views/AdmHeader.php';
         require 'app/views/AdmHome.php'; // View para a tela home do administrador
@@ -47,10 +51,15 @@ class AdmHomeController {
     
     // Método para excluir um aviso
     public function delete() {
-        // Certifique-se de que o ID está presente na URL
+        // Verifica se o usuário administrador está em sessão antes de permitir exclusão
+        if (!isset($_SESSION['admin_id'])) {
+            header('Location: /admlogin');
+            exit();
+        }
+
+        // Certifique-se de que o ID do aviso está presente na URL
         if (isset($_GET['id'])) {
             $avisoID = $_GET['id']; // Obtém o ID do aviso a ser excluído
-    
             $admModel = new AvisosAdmModel();
             $admModel->deleteAviso($avisoID);
         }
@@ -58,6 +67,4 @@ class AdmHomeController {
         header('Location: /homeadm'); // Redireciona para a tela home após a exclusão
         exit();
     }
-    
-    
 }
